extract render as module
